change system number version + new vue config + pwa experiment
