+++ README.md
@@ -7,15 +7,13 @@ Randomly generated field (5 by 5).
 Collect the number sequence from 1 to 25 strictly in order.
 Used vue.js and scss
 
+## version 0.1.1
+
+Решено изменить нумерацию версий
+Прошлые реализации в каталоге ./archive/
+Проект создан в учебных целях для изучения vue.js
+
 https://isschavliuga.ucoz.net
-
-# versions
-## <a href=https://codepen.io/ivanshavliuga/pen/pobKbpm>version 3.0.0</a>
-
-<iframe height="265" style="width: 100%;" scrolling="no" title="balls vue animate (crazy table shulte)" src="https://codepen.io/ivanshavliuga/embed/pobKbpm?height=265&theme-id=dark&default-tab=js,result" frameborder="no" loading="lazy" allowtransparency="true" allowfullscreen="true">
-  See the Pen <a href='https://codepen.io/ivanshavliuga/pen/pobKbpm'>balls vue animate (crazy table shulte)</a> by Ivan Shavliuga
-  (<a href='https://codepen.io/ivanshavliuga'>@ivanshavliuga</a>) on <a href='https://codepen.io'>CodePen</a>.
-</iframe>
 
 ``` bash
 cd iv2shulte3
@@ -31,28 +29,6 @@ npm run build
 
 # Lints and fixes files
 npm run lint
-```
-## version 2.0.0
-## Getting up and Running
-
-``` bash
-# Install the dependencies
-npm install
-
-# Serve with hot reload at http://localhost:8080
-npm run dev
-
-# Build for production with minification
-npm run build
-```
-
-## <a href=https://codepen.io/ivanshavliuga/pen/RqojWP>version 1.0.0</a>
-
-``` bash
-cd ver_1
-index.html
-#to resource
-https://codepen.io/ivanshavliuga/pen/RqojWP
 ```
 
 ## License
