+++ archive/ver0.1.0-webpack/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "tableshulte",
+  "description": "Table Shulte (Vue.js game)",
+  "version": "2.0.0",
+  "author": "Ivan Ivanov (Shavliuga)",
+  "email": "iva.drakon.nov@gmail.com",
+  "private": true,
+  "scripts": {
+    "dev": "cross-env NODE_ENV=development webpack-dev-server --open --inline --hot",
+    "build": "cross-env NODE_ENV=production webpack --progress --hide-modules"
+  },
+  "dependencies": {
+    "bootstrap": "^4.4.1",
+    "font-awesome": "^4.7.0",
+    "jquery": "^3.4.1",
+    "moment": "^2.17.1",
+    "vue": "^2.1.0",
+    "vuex": "^2.0.0",
+    "vue-router": "^2.0.0",
+    "scss": "^0.2.4",
+    "sass-loader": "^7.0.3",
+    "node-sass": "^4.9.2"
+  },
+  "devDependencies": {
+    "babel-core": "^6.0.0",
+    "babel-loader": "^6.0.0",
+    "babel-preset-es2015": "^6.0.0",
+    "cross-env": "^3.0.0",
+    "css-loader": "^0.25.0",
+    "file-loader": "^0.9.0",
+    "vue-loader": "^10.0.0",
+    "vue-template-compiler": "^2.1.0",
+    "webpack": "^2.1.0-beta.25",
+    "webpack-dev-server": "^2.1.0-beta.9"
+  }
+}
