+++ archive/ver0.1.0-webpack/src/main.js
@@ -1,0 +1,7 @@
+import Vue from 'vue';
+import App from './App.vue';
+
+let vm=new Vue({
+  el: '#app',
+  render: h => h(App)
+})
