+++ archive/ver0.1.0-webpack/src/timer.js
@@ -1,0 +1,19 @@
+export default {
+  current: {
+    hour:0,
+    minute:0,
+    second:0,
+    mark:'contol',
+    now(){
+      let d=new Date();
+      this.hour = d.getHours();
+      this.minute = d.getMinutes();
+      this.second = d.getSeconds();
+    }
+  }
+}
+/*window.onload = function() {
+current.now();
+setInterval(function() {current.now();}, 1000);
+};
+*/
