--- iv2shulte3/README.md
@@ -1,38 +1,0 @@
-# iv2shulte3
-## Таблиа Шульте
-Генерируется случайным образом поле (5 на 5).
-Собирайте числовую последовательность от 1 до 25 строго по порядку.
-Используется vue.js и scss
-
-Randomly generated field (5 by 5).
-Collect the number sequence from 1 to 25 strictly in order.
-Used vue.js and scss
-
-https://isschavliuga.ucoz.net
-
-
-## Project setup
-```
-npm install
-```
-
-### Compiles and hot-reloads for development
-```
-npm run serve
-```
-
-### Compiles and minifies for production
-```
-npm run build
-```
-
-### Lints and fixes files
-```
-npm run lint
-```
-
-### Customize configuration
-See [Configuration Reference](https://cli.vuejs.org/config/).
-
-## License
-MIT
