+++ public/manifest.json
@@ -1,0 +1,20 @@
+{
+  "name": "vuemp3player",
+  "short_name": "vuemp3player",
+  "icons": [
+    {
+      "src": "./img/icons/iv2-128x128.png",
+      "sizes": "128x128",
+      "type": "image/png"
+    },
+    {
+      "src": "./img/icons/iv2-512x512.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    }
+  ],
+  "start_url": "./index.html",
+  "display": "fullscreen",
+  "background_color": "#7e22a1",
+  "theme_color": "#7e22a1"
+}
