+++ src/main.js
@@ -1,7 +1,13 @@
-import Vue from 'vue';
-import App from './App.vue';
+import Vue from 'vue'
+import App from './App.vue'
+import router from './router'
+import store from './store'
+import './registerServiceWorker'
 
-let vm=new Vue({
-  el: '#app',
+Vue.config.productionTip = false
+
+new Vue({
+  router,
+  store,
   render: h => h(App)
-})
+}).$mount('#app')
